Removed dependency on LoggerService because although its part of the OSGi framework, it would make it hard to unit test these classes outside a container, when a simple slf4j logger is sufficient.
Removed the multitude of constructors probably left over from the guice version of this set of classes. I would prefer that
we do constructor IOC, but while OSGi cant to this we are going to have to live with lots of unnecessary getter and setter code,
extra code for make the classes thread safe on construction and potentially extra code to make certain that the classes are in an
initalized state. This all makes my heart sink, more unnecessary complexity. :(

Modified the bundle plugin configuration, you should not really do selective imports, you should import * and exclude the things that you
dont need. The otherway round results in unknow missing dependencies that appear at runtime as either ClassNotFoundException or NoClassDefError.
The latter are hard to diagnose and could happen at any time.