KERN-139, KERN-144
Added listing with paging and sorting of site members. At the moment this uses in memory sorts because all the data comes
from JCR which is already in memory so the iterators do no cause data base traffic. However I am not entirely comfortable
with this. We may need to push the page and sort back to a JPA table, but the cost of maintianing a published hierachical tree
flattened for every group is going to be expensive as each external node validation will cause a republication of every node the
group or user is connected to. For some users thats the entire group space.

Following the google model, where there are more than 10K users per site the system will not list. A warning will be emmitted and no
results will be returned. We may also need to cache the flattened trees per site, but there is a simple place to do that.