Ported the ResourcePath* Resolvers to a Single VirtualResourceProvider that listens for VirtualResourceType implementations.
This eliminates the need for API changes in Sling but keeps the processing the same as it is now. At the moment the VirtualResourceType implementations emit synthetic resources, but we might change this in future to emit JcrNodeResources. Its up to the servlet to manage how it handles the Synthetic Resource. It might also make sense to use a better VirtualResource at this point.