Fixed login problem
Sling encodes passwords before sending them to the user manager for comparison
Sling also encodes passwors on creation before sending them to the user manager for comparison

The Authenticator in Sling, seperately encodes passwords

Although we encode passwords in the Sparse code, we do it all in one place, and use SHA-512 all the time, this is
to ensure 100% that a password can never be stored plain text in the DB. So when interfacing to Sling, as it scatters the
password encryption we have to try first the password digested by sling, and then by SHA-512, if that fails we try
digesting the password with SHA-512 raw. This allows us to use a single password authenticator for all authentication
against the sparse database.