KERN-1501 Fixed build, startup Ok, sorry for any inconvinence and delay.
Note, this sligtly changes the way connected user works. Rather than using an admin Session it will use
the current request session or supplied session, so that any ACLs on the usr connections is honored.
If a connection is made private to a subset of users, only those users will be able to see the connection.
Before any user could have seen the connections.