SAKIII-2252 This fix includes 2 particular bits: 1) Updates to our apache sakai23hybrid.conf to rewrite our /portal/site/{siteid} urls to the corresponding OAE Hybrid views. It's likely these rewrites will be continue to be revisited as we debug, expand, and test Hybrid functionality. Ideally, they are generic enough to also be implemented in other systems as well, such as future native OAE Nakamura Proxy functionality, nginx rules, Zues rules, or other frontend/load balancers.
2) Additional handing in our OAE Hybrid view to allow navigating
straight to a particular page in a site with the addition of a
pageid query parameter. This portion of the fix also moves some of
CSS class handling for loading Sakai2Page Views into the loadPageTools
function so that function can act as a single location for swapping
the views (rather than just the click handler).