KERN-1939  When performing MoreLikeThis queries, ignore any fields that don't have term vectors
When Solr is asked to calculate similar documents based on a field that doesn't have term vectors, it tokenises the stored value of that field (if present) at runtime.  For fields with many values, this adds a lot of overhead to each query and leads to slow queries.

By disabling this behaviour, fields used in MoreLikeThis calculations will need to have term vectors enabled or will be silently ignored.  This isn't ideal, but seems like a better outcome than sudden hard to explain query performance problems.