NOJIRA Added a new FileUploadHandler interface for hooking into the file upload process.
NYU has developed a bundle to integrate video content uploaded to Nakamura with the Kaltura video platform.  This required an additional hook to be added to the CreateContentPoolServlet to let bundles provide listeners for uploaded files.  This patch adds a FileUploadHandler interface, and implementing classes provide a handleFile() method which is called when a new file is uploaded, allowing them to act on the file's binary content and metadata as required.