KERN-1932 Modified SolrResultSetFactory to listen for content deletion events and exclude deleted paths from search results.
Deleted paths are stored until the ContentEventListener signals (via a org/sakaiproject/nakamura/solr/COMMIT event) that the index has been committed.

To give this a chance at working in a clustered environment, deleted paths are stored by using the memory service as a sort of shared memory.  Each node writes the deleted paths it knows about to the shared cache, and the global list of shared paths is taken into account when limiting search results.